Revert "Refactor/lecturer listeners (#371)"
This reverts commit 3b165d2d966774da4a6ef3a88f1d52d8e477b214.
