+++ src/lecturer/components/reactionReceiveView/ReactionReceiveView.tsx
@@ -3,14 +3,12 @@ import React, { useContext, useCallback, useEffect, useState } from "react";
 import { StoreContext } from "../../services/StoreService";
 import { ReactionCounter } from "./ReactionCounter";
 import { reactionsIcons } from "../../../common/util/reactions/icons";
+import { useSocket } from "../../services/SocketService";
 import { ReactionName } from "../../../common/util/reactions/enum";
 import { ReactionProgress } from "./ReactionProgress";
-import { useSocket } from "../../services/SocketService";
 
 export function ReactionReceiveView() {
     const store = useContext(StoreContext);
-    const { socketEmiter } = useSocket();
-
     const reactions = [
         ReactionName.HEART,
         ReactionName.HAPPY,
@@ -18,6 +16,32 @@ export function ReactionReceiveView() {
         ReactionName.UP,
         ReactionName.DOWN,
     ];
+    // hotfix by me
+    // TODO get reaction index with string sent with payload
+    // TODO make it work nice and properly, because I do not know how 
+    const reactionsString = [
+        "HEART",
+        "HAPPY",
+        "SAD",
+        "UP",
+        "DOWN"
+    ];
+    const TIME_WAITING = 20000;
+    const { socketEmiter } = useSocket();
+    const refreshReactions = useCallback((payload?: ReactionResponsePayload) => {
+        let index: number;
+        if (payload) {
+            let indexString: string = payload.data.reaction;
+            index = reactionsString.indexOf(indexString);
+        } else {
+            index = Math.round(Math.random() * reactionsString.length);
+        }
+        let tmpValues = store.reactionValues;
+        tmpValues[index]++;
+        store.reactionValues = tmpValues;
+        if (!store.reactionModes[index] || store.lastReactionTime > 0)
+            store.lastReactionTime = Date.now() + TIME_WAITING;
+    },[]);
     
     const [progressEnabled, setProgressEnabled] = useState<boolean>(!store.reactionModes.reduce((acc, mode) => {
         return acc && mode;
@@ -65,6 +89,13 @@ export function ReactionReceiveView() {
         return () => clearInterval(interval);
     }, [store, store.lastReactionTime, resetReactions]);
 
+    useEffect(() => {
+        socketEmiter.on("send_student_reaction", refreshReactions);
+        return () => {
+            socketEmiter.off("send_student_reaction", refreshReactions);
+        };
+    }, [refreshReactions, socketEmiter]);
+
     const classes = makeStyles({
         reactionWrapper: {
             padding: 10,
@@ -74,7 +105,7 @@ export function ReactionReceiveView() {
             overflow: "hidden",
         },
     })();
-    const TIME_WAITING = 20000;
+
     return (
         <Paper variant="outlined" square>
             {progressEnabled && <ReactionProgress time={TIME_WAITING} />}
